Fix compile error on ArduinoIDE
refs #7431

+++ Arduino15/packages/SPRESENSE/hardware/spresense/1.0.0/libraries/DNNRT/DNNRT.cpp
@@ -105,7 +105,7 @@ DNNRT::end()
 int
 DNNRT::inputVariable(DNNVariable var, unsigned int index)
 {
-  if (index >= _nr_inputs)
+  if (index >= (unsigned int)_nr_inputs)
     {
       return -1;
     }
